requirements.txt
asgiref==3.7.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
astroid==3.0.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
asttokens==2.4.1 ; python_version >= "3.11" and python_version < "3.12.dev0"
black==24.4.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
boto3==1.34.84 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
botocore==1.34.84 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
certifi==2024.2.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
cfgv==3.4.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
charset-normalizer==3.3.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
click==8.1.7 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
colorama==0.4.6 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0" and (sys_platform == "win32" or platform_system == "Windows")
decorator==5.1.1 ; python_version >= "3.11" and python_version < "3.12.dev0"
dill==0.3.8 ; python_version >= "3.11" and python_version < "3.12.dev0"
distlib==0.3.8 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-anymail==10.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-ckeditor-5==0.2.12 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-ckeditor==6.7.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-environ==0.11.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-extensions==3.2.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-filter==24.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-js-asset==2.2.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-phonenumber-field[phonenumbers]==7.3.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-simple-history==3.5.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-storages==1.14.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-viewflow==2.0.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-webpack-loader==3.1.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django==5.0.4 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
djangorestframework-camel-case==1.4.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
djangorestframework==3.15.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
drf-base64==2.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
executing==2.0.1 ; python_version >= "3.11" and python_version < "3.12.dev0"
factory-boy==3.3.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
faker==23.2.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
filelock==3.13.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
identify==2.5.35 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
idna==3.7 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
ipdb==0.13.13 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
ipython==8.21.0 ; python_version >= "3.11" and python_version < "3.12.dev0"
isort==5.13.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
jedi==0.19.1 ; python_version >= "3.11" and python_version < "3.12.dev0"
jmespath==1.0.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
matplotlib-inline==0.1.6 ; python_version >= "3.11" and python_version < "3.12.dev0"
mccabe==0.7.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
mypy-extensions==1.0.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
nodeenv==1.8.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
packaging==23.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
parso==0.8.3 ; python_version >= "3.11" and python_version < "3.12.dev0"
pathspec==0.12.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pexpect==4.9.0 ; python_version >= "3.11" and python_version < "3.12.dev0" and sys_platform != "win32"
phonenumbers==8.13.33 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pillow==10.3.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
platformdirs==4.2.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pre-commit==3.7.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
prompt-toolkit==3.0.43 ; python_version >= "3.11" and python_version < "3.12.dev0"
psycopg2==2.9.9 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
ptyprocess==0.7.0 ; python_version >= "3.11" and python_version < "3.12.dev0" and sys_platform != "win32"
pure-eval==0.2.2 ; python_version >= "3.11" and python_version < "3.12.dev0"
pygments==2.17.2 ; python_version >= "3.11" and python_version < "3.12.dev0"
pylint-django==2.5.5 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pylint-plugin-utils==0.8.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pylint==3.0.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
python-dateutil==2.8.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pyyaml==6.0.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
requests-mock==1.12.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
requests==2.31.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
s3transfer==0.10.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
sentry-sdk==1.45.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
setuptools==69.1.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
sib-api-v3-sdk==7.6.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
six==1.16.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
sqlparse==0.5.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
stack-data==0.6.3 ; python_version >= "3.11" and python_version < "3.12.dev0"
tomlkit==0.12.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
traitlets==5.14.1 ; python_version >= "3.11" and python_version < "3.12.dev0"
tzdata==2024.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0" and sys_platform == "win32"
urllib3==2.0.7 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
virtualenv==20.25.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
wcwidth==0.2.13 ; python_version >= "3.11" and python_version < "3.12.dev0"
